requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
comtypes==1.1.7
decorator==4.4.2
dragonfly2==0.28.1
idna==2.10
json-rpc==1.13.0
kaldi-active-grammar==1.8.1
lark-parser==0.8.9
numpy==1.19.3
packaging==20.8
pycparser==2.20
pyparsing==2.4.7
pyperclip==1.8.1
pywin32==300
regex==2020.11.13
requests==2.25.1
six==1.15.0
sounddevice==0.3.15
urllib3==1.26.2
ush==3.1.0
webrtcvad-wheels==2.0.10.post2
Werkzeug==1.0.1
